Use reusable venvs
Replace pip install --target with venvs.
As setuptools breakes out of the venv to load distutils build_ext
command, first create a copy of the Python distribution and hack
it.

This also allows venv's to be populated manually outside of
PyOxidizer's config, while retaining the ability for PyOxidizer
to pick up and include any built object files.

To facilitate that, the PyOxidizer state dir where objects are
stored is a constant in the distutils compiler modules and in
the Rust source.

Also add venv_path to PipRequirementsFile, allowing the same
venv to be incrementally populated in multiple rules,
and the venv re-used across PyOxidizer build runs.

Fixes https://github.com/indygreg/PyOxidizer/issues/162
Fixes https://github.com/indygreg/PyOxidizer/issues/170
Closes https://github.com/indygreg/PyOxidizer/issues/194